Rust SDK: remove `anyhow`, use `thiserror` w/ structured error types
Our use of `anyhow` in the Rust SDK was originally due to my being lazy.
I figured I'd kick the can on writing actual error types as long as possible,
and it turns out now's the time that's no longer possible,
because while working on a follow-up I needed to `clone` an error,
and `anyhow::Error: !Clone`.

This commit removes the Rust SDK's use of `anyhow`,
replacing it with a `thiserror`-annotated `enum Error`.

This resulted in essentially simple mechanical changes to  a lot of the SDK code,
and also to the codegen, resulting in a truly enormous diff.