Rust client SDK: rework `EventContext` into multiple types
A repeated pain point with the post-0.12 client SDK API
has been excessive unnecessary pattern-matching or unwrapping
on the `ctx.event` within reducer callbacks.
Previously, these had `ctx.event: Event`,
with the type system not knowing that said `Event` would always be `Event::Reducer`.
This was in a misguided attempt to reduce proliferation of context types.

As of this PR, we have different event context types for each category of callback,
which allows us to inform the type system precisely of the type of `event`.

The event context types:

- EventContext, passed to row callbacks.
  - Exactly the previous type.
  - In the future we may examine enum Event more closely and see if any of its variants are unreachable in row callbacks. If they are, we can deprecate those variants.
- ReducerEventContext, passed to reducer callbacks.
  - `event: ReducerEvent<Reducer>`.
- SubscriptionEventContext, passed to subscription applied and unapplied callbacks.
  - No event field needed, or `event: ()`.
- ErrorContext, passed to various on-error callbacks.
  - `event: Error`.

Changes to callbacks:

- `DbConnectionBuilder::on_disconnect` now takes `FnOnce(&ErrorContext)`.
  - For "normal" disconnects, this will have `event: Error::Disconnected`.
  - Is this better than having a separate `DisconnectContext`, which would likely also hold `event: Error`?
- `DbConnectionBuilder::on_connect_error` now takes `FnOnce(&ErrorContext)`.
- `on_{reducer}` now takes `FnMut(&ReducerEventContext, ...Args)`.
- `SubscriptionBuilder::on_applied` and `SubscriptionHandle::unsubscribe_then` now take `FnOnce(&SubscriptionEventContext)`.
- `SubscriptionBuilder::on_error` now takes `FnOnce(&ErrorContext)`.