Remove more `Address` references
Incl. two files that were not reachable and seemed severely bit-rotted:
- `tracelog.rs`
- `reducer.rs`

The latter I myself forgot to delete in a past PR;
the former I do not know about, but it's not part of our project,
it wouldn't compile, and I don't see any reason to keep it in our repo.