sdk: Basic connection metrics
Instruments the websocket loop to track number and size of received
messages. Mirrors the metrics we have on the sending side.

Motivated by having observed clients that permanently return stale data,
even though both ends of the connection appear to consider the
connection healthy.

There may be other interesting metrics (such as the size of the client
cache), but there is no immediate need, so we're deferring this to a
future addition.

Another future change may be to actively terminate the connection if no
data (including PING frames) is received within a timeout.