commitlog: Fix open flags for read-only offset index
For some reason, we've been opening the read-only index file with
`.read(true).append(true)`.

This translates to `O_RDWR | O_APPEND` on *nix, but something weird on
Windows. A shared memory map requires `O_RDWR` (or equivalent) on either
platform, though, meaning that we'd always fail to `mmap` the index on
Windows.

Open with `.read(true).write(true)` instead to fix.