Various methods and trait impls for time arithmetic
This commit adds `checked_add` and `checked_sub` methods
for arithmetic with `Timestamp` and `TimeDuration`,
and implementations of `Add`, `Sub`, `AddAssign` and `SubAssign`.

Specifically, we add methods and trait impls for `T x U -> V`, where `x` is a binary operator,
either + or -:

- `Timestamp x TimeDuration -> Timestamp`
- `Timestamp x Duration -> Timestamp`.
  - This one is kind of weird, since we convert the `Duration` into a `TimeDuration`
    before doing math with it, which may be a lossy conversion depending on platform.
- `TimeDuration x TimeDuration -> TimeDuration`.

These correspond to the operations that Rust exposes on
`std::time::SystemTime` and `std::time::Duration`,
plus allowing `Duration` as a convenience.

I have chosen not to implement
`TimeDuration x Duration` or `Duration x TimeDuration` operations,
under the theory that if you're already working with a `TimeDuration`,
you don't need to be shielded from the conversion.