Terminate connections when `client_connected` reducer returns error
Prior to this commit, if a module defined a `client_connected` reducer,
and that reducer returned an error for a connection to a database,
the connection would still be allowed to proceed.
With this commit, instead, the client's connection will be terminated.
This allows module developers to implement various authorization logic.

This commit also adds a smoketest, `client_connected_error_rejects_connection`,
which verifies that returning an error from `client_connected`
prevents the connection from proceeding.

Also, we refactor `call_identity_connected_disconnected` into two separate methods,
`call_identity_connected` and `call_identity_disconnected`,
because the two now have meaningfully different behavior.