Repair deleting from `st_client` if `client_disconnected` errors
It turned out to be the same problem as we had in the `client_connected` path:
`call_reducer_inner` looks like it should return `Err` if the reducer errored,
but it doesn't. The authors of this method, including myself when I rewrote it yesterday,
understandably but misguidedly assumed that an `Ok` result from `call_reducer_inner`
meant that the transaction had committed, and so didn't check the `ReducerOutcome`.