Rip Rayon out of reducer execution and query evaluation
For testing/benchmarking purposes.
One theory about our performance is that we're spending a lot of time context-switching
between Tokio and Rayon threads.
This build will be used in the first of a series of experiments
to investigate that overhead.
In this patch, we just do sequential evaluation on the Tokio worker threads
where the rest of our code runs, instead of sending stuff to Rayon.
Rayon use is mostly, but not entirely, removed from Core.

The two next steps I am interested in:
- Use parallelism, but on Tokio workers rather than Rayon threads.
  I.e. replace `par_iter`, `fold` and `reduce_with` calls with Tokio-isms,
  instead of this patch's `std::iter::Iterator` versions.
- (Not discussed in meeting) continue using `par_iter` and friends,
  but invoke the "outer loop" from Tokio threads.
  I.e. retain use of `par_iter`, `fold` and `reduce_with`,
  but remove calls to `rayon::scope` or `rayon::spawn`.