Roll back control DB modifications when failing to schedule replicas
The only way we have to determine whether a module is valid
is to construct a database and at least one replica of it.
This is unfortunate, since our control DB in standalone is not transactional.
So we just check if scheduling replicas failed during publish,
and if it did, we roll back the control DB insert.

This doesn't occur when publishing onto an existing database,
since the automigration executor checks the new RLS rules,
and it gracefully handles failure.