Impose limits on the heap usage of `ChunkPool`
Heap profiling performed by @jsdt revealed that we had a memory leak in
`InstanceEnv::datastore_index_scan_range_bsatn_chunks`.
A bit of spelunking pointed to `ChunkPool`, a pool of buffers,
which prior to this commit was allowed to grow without limit.
`ChunkPool` is per-database, and so it is easy to imagine a scenario in which
a large number of databases each occasionally
allocate, use and never free very large chunks,
resulting in the overall system's memory usage increasing over time.

This commit adds some very simple limits on the heap usage of a `ChunkPool`.
Specifically, we introduce two constants,
`MAX_CHUNKS_IN_POOL` and `MAX_CHUNK_SIZE_IN_BYTES`,
and when returning a chunk to the pool, we free it if it violates these limits.
The values of these constants, 32 and 1024 respectively, were chosen entirely arbitrarily.
We should experiment at least to demonstrate that they are not pathologically bad,
though I doubt we actually care to optimize them beyond that.