Introduce a `DatabaseLifecycleTracker`, which tracks database lifecycles
Perhaps it should be called `DatabaseLifecycleManager`?

This new object is responsible for tracking the lifecycle of a database,
and for cleaning up after the database exits.
In particular, it:

- Unregisters the `Host` from the containing `HostController`.
  This was previously handled by an ad-hoc on-panic callback closure.

- Aborts the database memory usage metrics reporter task.
  This was previously handled by a `Drop` method on `Host`.

- Disconnects all connected WebSocket clients.
  Previously, this didn't happen at all, as per issue 2630.

I've also added some commentary to the WebSocket actor loop.

Follow-up commits will add tests once I've consulted with the team
about how best to test this change.