mTLS + some fixes(1MiB thing was broken badly)
mTLS=mutual TLS, client needs to send cert to be authenticated by the
server
mTLS is currently optional, depending on if the client args for client certs
are specified. Everything should work as before otherwise.

Improved the description of some connection errors, like when client cert
is required but you didn't specify the file args for it.