Fix SDK bug with overlapping subscriptions
Prior to this commit,
the Rust client SDK was dropping all `TableUpdate`s other than the last
within a given `DatabaseUpdate`,
leading to incorrect client cache states, particularly incorrect row multiplicities.
This was reported by @lakmatiol in the public Discord,
who was kind enough to provide us with a minimal reproducer,
https://github.com/lavirlifiliol/spacetime-repro .

This commit fixes the bug by combining `TableUpdate`s rather than overwriting them.
It also adds a new SDK test which does the same thing as the provided reproducer.

Note that this fix affects both the codegen and the SDK itself,
meaning we should hold it until a minor version bump.