Prevent `ahash` from pulling in `getrandom 0.3`
Modules generally don't honor the workspace cargo lock file, so may
upgrade `ahash` silently, which pulls in a `getrandom` version that is
not the one pulled in via `binding`. Thus feature unification doesn't
apply and modules fail to compile due to the wasm target not being
supported by the transitive `getrandom`.

This also causes smoketests to fail.