ci: Update dependencies for the `wasm_bindings` job
Make sure the `Cargo.lock` file of the `wasm_bindings` job reflects the
latest available versions. This is what users would end up with on a
fresh module, so we want to catch any compile errors arising from a
different transitive closure of dependencies than what is in the
workspace lock file.

Follup-up for: https://github.com/clockworklabs/SpacetimeDB/pull/2714