Reintroduce replication factor parameter
The `--num-replicas` option was removed from the CLI pre-1.0
It is re-introduced here as a hidden / unstable option to facilitate
testing (e.g. to override defaults).

The `DatabaseDef` type is changed to make `num_replicas` optional, so as
to distinguish "use the default" (`None`) from "shut down all replicas
but keep the database" (`Some(0)`).