Reject clients who match an existing `st_client` row
At various places in our codebase, we assume that client connections
are unique by `(identity, connection_id)` pair,
but we never actually enforced this prior to now.
It was difficult, but not impossible, to accidentally connect
with an already-present `(identity, connection_id)` pair.
The easiest way was to open two `DbConnection`s in parallel
from within a single Rust client SDK process,
due to a misbehavior in that SDK which I intend to fix in a separate PR.

This commit modifies `MutTxId::insert_st_client` to return an error
if the row to be inserted is already resident in the database.
It's still possible for a database owner to delete from `st_client` manually
via a SQL `delete` statement, which will cause strange misbehaviors,
but it should no longer be possible for an unprivileged client
to accidentally put the database into a state that's intended to be impossible.