Add `client_connection_incoming_queue_length` metric
It's a metric that tracks the size of the incoming per-client `message_queue`.
We've been concerned about not having visibility into this queue's length,
as currently we only have the length of the per-database reducer queue,
but each client can only have a single reducer in that queue,
and may have additional messages waiting in its per-client queue.

The new metric, `spacetime_client_connection_incoming_queue_length`,
is an `IntGaugeVec` with the labels:
`db: Identity, client_identity: Identity, connection_id: ConnectionId`.
My theory is that in our viewer we can inspect the average and the sum per database,
and it also may be interesting to be able to look at individual clients,
as e.g. the BitCraft mob monitor may be a notable outlier.
This is not the same pattern as most of our other metrics, though,
which tend to only offer per-database granularity.
It's possible that this new metric should also have the last two labels removed,
and be labeled only on `db: Identity`.