`build_sequence_state`: clobber pre-existing in-memory `Sequence`s
Fixes #2758 .

Prior to this commit, system sequences (as in, sequences for columns of system tables)
were not being correctly re-initialized after replaying from a commitlog.
The reproducer for this was to delete the zero snapshot and then replay,
and then to advance the sequence  by publishing a new module version with a new table.

The difference which hid this bug was that, specifically when replaying from nothing,
we begin by calling `Locking::bootstrap`, which will `build_sequence_state`,
and then after replaying the commitlog,
we `rebuild_state_after_replay`, which will `build_sequence_state` again.
When replaying from a snapshot, we bypass `bootstrap`,
and so the only call to `build_sequence_state` is in `rebuild_state_after_replay`
after replaying the suffix of the commitlog.

The problem was that, prior to this commit,
`build_sequence_state` did not fully overwrite the in-memory sequence state.
If an in-memory `Sequence` already existed in the `SequencesState`,
`build_sequence_state` would reset that sequence's `value`, but not its `allocation`.
Such a pre-existing in-memory `Sequence` could only exist (1) for system sequences
and (2) when replaying without a snapshot.
If either of these conditions were not met, no `Sequence` object would yet exist,
and so `build_sequence_state` would install all of the values from the on-disk state.

The fix is simple: unconditionally clobber the in-memory `Sequence` state
during `build_sequence_state`, always using the values from the committed state.