Less granular message queue length metrics
This commit alters the message queue length metrics
introduced by https://github.com/clockworklabs/SpacetimeDB/pull/2754
to be per-database, rather than per-client.
This should limit the cardinality of these metrics,
and better lines up with the labels of our other metrics.

Because the metrics are now per-database rather than per-client,
it's no longer correct to just drop the label when the client disconnects.
Instead, care must be taken to decrement the metric
by the number of messages which were waiting in the queue
at the time of the disconnection.
I've added comments to call attention to this complexity.