Add metric for queue length
This commit adds a new metric, `subscription_send_queue_length`.
It is incremented by `eval_updates_sequential` before pushing into the new queue,
and decremented by the `send_worker` after popping.

Because the metric is labeled with `database_identity: Identity`,
the `SubscriptionManager` now needs to know its `Identity`, and so cannot be `Default`.
Methods are added to make constructing one in tests easier.