core: Allow adjusting CPU reservation pools
Introduces a way to reserve another fixed amount from the available
cores, and allows to access the `CpuSet` used for tokio blocking
threads.

This is motivated by wanting to create a separate tokio runtime that can
share the blocking pool.

The patch also allows to adjust the fractional values for the various
pools, and exports methods for inspecting the actual assignments.