Enable the `datastore` crate's `test` feature where necessary
2933 broke running `cargo test` without `--all`
because the new `datastore` crate hid some of its test utilities
behind `cfg(any(test, feature = "test"))`,
but did not enable that feature in the dependency from `core` or `sqltest`.

This was hidden in our CI because we run with `--all` there,
which apparently enables all optional features, including the `test` feature.