refactor(sdk): improve the log message source line reference (#2924)
# Description of Changes

Currently with the function maybe_log_error to log errors we hide the
real source line where the error happened. So to still have the same
logic but with more accurate source line, I moved it to a macro to have
the same abstraction for logging.

# API and ABI breaking changes

- 

# Expected complexity level and risk

1

# Testing

- [ ] I have tested it my application and it now show the better source
line of the log message

---------

Signed-off-by: ResuBaka <ResuBaka@users.noreply.github.com>
Co-authored-by: Phoebe Goldman <phoebe@goldman-tribe.org>