Increase the default incoming-queue-length limit
2048 turned out to be too low a value for BitCraft,
as their world upload process requests on the order of 6000 reducers very rapidly.
We still feel that having a limit is valuable to prevent malicious or misguided clients
from taking an arbitrarily large amount of host memory,
so we bump the value to give us a wide safety error for BitCraft's needs
but don't remove the limit entirely.