Sort columns in `st_column_changed` before automigrating.
During bootstrapping, indexes don't exist,
so `iter_by_col_eq` returns rows in physical storage order.
In simple cases (without unrelated deletes) this will be the same as insertion order,
and therefore also the same as sorted order,
but in cases with multiple column-changing automigrations
the physical storage order of the rows diverges from the correct sorted order.
Because this isn't a hot path, we can just call `.sort()` and not worry about it.