Removed // @ts-ignore directive from generated files and fixes associated errors (#3228)
# Description of Changes

Previously all TypeScript generated code had a `// @ts-ignore` directive
at the top to suppress unused variable errors that users may have run
into with their generated code. This worked to suppress unused, but it
was too broad and suppressed actual errors with our generated code. This
PR removes that directive and fixes the latent errors.

Most notably this fixes the issue in the quickstart where we were
getting `serialize` is not a function.

# API and ABI breaking changes

None

# Expected complexity level and risk

1

# Testing

<!-- Describe any testing you've done, and any testing you'd like your
reviewers to do,
so that you're confident that all the changes work as expected! -->

- [x] This PR increases the test surface only by now typechecking the
typescript generated code within our repository